requirements.txt
streamlit>=1.32,<2.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
python-dateutil>=2.9
supabase>=2.4,<3.0
